huge responsiveness added bootstrap manipulated for 8 different route, ready to test current functionality for all routes
