+++ client/src/pages/MainPage.jsx
@@ -10,7 +10,7 @@ import {
   OverlayTrigger,
   Accordion,
 } from "react-bootstrap";
-import { LogoutModal, PostModal, CommentModal } from "../components/ui/Modal";
+import { CommentModal } from "../components/ui/Modal";
 
 import { Link, useNavigate } from "react-router-dom";
 import axios from "axios";
@@ -33,6 +33,8 @@ const API_URL = "http://localhost:3000";
 // when working on deployment version
 // ?
 import io from "socket.io-client";
+import LeftSideNavBar from "../components/Main-Left-Side-Navbar/LeftSideNavbar";
+import RightSideColumn from "../components/Main-Right-Side-Column/RightSideColumn";
 
 function MainPage() {
   const socket = io.connect(`${API_URL}`);
@@ -48,7 +50,9 @@ function MainPage() {
     const getClickLocation = (e) => {
       const clickedElementParentClass = e.target.parentNode.className;
       const clickedElementClass = e.target.classList;
-
+      console.log("Parent class =>", clickedElementParentClass);
+      console.log("Child class =>", clickedElementClass);
+      console.log("Clicked post =>", clickedPostBox);
       if (
         (clickedElementClass.contains("hover-reposted-text") &&
           clickedElementParentClass !== "post-circle-profile-svg-on-point" &&
@@ -70,7 +74,10 @@ function MainPage() {
           "outside-of-inner-circle-post-info-user-info-svg-three-dots hstack gap-1" ||
         clickedElementParentClass === "mt-0 parent-footer-stack hstack" ||
         clickedElementClass.contains("repost-svg-post-box") ||
-        clickedElementParentClass === "post-head"
+        clickedElementParentClass === "post-head" ||
+        clickedElementClass.contains("each-post") ||
+        clickedElementClass.contains("border-extra") ||
+        clickedElementParentClass === "each-post"
       ) {
         if (clickedPostBox) {
           navigate(
@@ -618,13 +625,28 @@ function MainPage() {
     setvisibleFollowingTweets((prevVisibleTweets) => prevVisibleTweets + 25);
   };
 
+  const [windowWidth, setWindowWidth] = useState(window.innerWidth);
+
+  console.log("Current inner width =>", windowWidth);
+
+  const handleResize = () => {
+    setWindowWidth(window.innerWidth);
+  };
+
+  useEffect(() => {
+    window.addEventListener("resize", handleResize);
+
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
   return (
     <>
       {contextHolder}
       <ToastContainer />
       <ResponsiveNavigationBarBottom />
       <ResponsiveNavigationBarTop />
-      <Container>
+      <Container fluid>
         <Row
           style={{
             height: "100vh",
@@ -632,128 +654,36 @@ function MainPage() {
             borderBottom: "none",
           }}
         >
-          <Col className="left-column" xs={12} sm={12} md={1} lg={3} xxl={3}>
-            <nav className="nav-bar-home">
-              <Link href="/home">
-                <div className="chevron-left-parent">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="30"
-                    height="30"
-                    fill="currentColor"
-                    className="bi bi-chevron-double-left like-icon"
-                    viewBox="0 0 20 20"
-                  >
-                    <path d="M8.354 1.646a.5.5 0 0 1 0 .708L2.707 8l5.647 5.646a.5.5 0 0 1-.708.708l-6-6a.5.5 0 0 1 0-.708l6-6a.5.5 0 0 1 .708 0z" />
-                    <path d="M12.354 1.646a.5.5 0 0 1 0 .708L6.707 8l5.647 5.646a.5.5 0 0 1-.708.708l-6-6a.5.5 0 0 1 0-.708l6-6a.5.5 0 0 1 .708 0z" />
-                  </svg>
-                </div>
-              </Link>
-
-              <div className="inner-div-fonts inner-div">
-                <Link to="/home">
-                  <div className="home">
-                    <div style={{}}>
-                      <svg
-                        width={26}
-                        height={26}
-                        viewBox="0 0 24 24"
-                        aria-hidden="true"
-                        className=" r-4qtqp9 r-yyyyoo r-dnmrzs r-bnwqim r-1plcrui r-lrvibr r-18jsvk2 r-lwhw9o r-cnnz9e"
-                      >
-                        <g>
-                          <path d="M21.591 7.146L12.52 1.157c-.316-.21-.724-.21-1.04 0l-9.071 5.99c-.26.173-.409.456-.409.757v13.183c0 .502.418.913.929.913H9.14c.51 0 .929-.41.929-.913v-7.075h3.909v7.075c0 .502.417.913.928.913h6.165c.511 0 .929-.41.929-.913V7.904c0-.301-.158-.584-.408-.758z"></path>
-                        </g>
-                      </svg>
-
-                      <span>Home</span>
-                    </div>
-                  </div>
-                </Link>
-
-                <Link>
-                  <div className="notifications">
-                    <div>
-                      <svg
-                        width={26}
-                        height={26}
-                        viewBox="0 0 24 24"
-                        aria-hidden="true"
-                        className="r-4qtqp9 r-yyyyoo r-dnmrzs r-bnwqim r-1plcrui r-lrvibr r-18jsvk2 r-lwhw9o r-cnnz9e"
-                      >
-                        <g>
-                          <path d="M19.993 9.042C19.48 5.017 16.054 2 11.996 2s-7.49 3.021-7.999 7.051L2.866 18H7.1c.463 2.282 2.481 4 4.9 4s4.437-1.718 4.9-4h4.236l-1.143-8.958zM12 20c-1.306 0-2.417-.835-2.829-2h5.658c-.412 1.165-1.523 2-2.829 2zm-6.866-4l.847-6.698C6.364 6.272 8.941 4 11.996 4s5.627 2.268 6.013 5.295L18.864 16H5.134z"></path>
-                        </g>
-                      </svg>
-
-                      <span>Notifications </span>
-                    </div>
-                  </div>
-                </Link>
-
-                <Link to="/messages">
-                  <div className="messages">
-                    <div>
-                      <svg
-                        width={26}
-                        height={26}
-                        viewBox="0 0 24 24"
-                        aria-hidden="true"
-                        className="r-4qtqp9 r-yyyyoo r-dnmrzs r-bnwqim r-1plcrui r-lrvibr r-18jsvk2 r-lwhw9o r-cnnz9e"
-                      >
-                        <g>
-                          <path d="M1.998 5.5c0-1.381 1.119-2.5 2.5-2.5h15c1.381 0 2.5 1.119 2.5 2.5v13c0 1.381-1.119 2.5-2.5 2.5h-15c-1.381 0-2.5-1.119-2.5-2.5v-13zm2.5-.5c-.276 0-.5.224-.5.5v2.764l8 3.638 8-3.636V5.5c0-.276-.224-.5-.5-.5h-15zm15.5 5.463l-8 3.636-8-3.638V18.5c0 .276.224.5.5.5h15c.276 0 .5-.224.5-.5v-8.037z"></path>
-                        </g>
-                      </svg>
-                      <span>Messages</span>
-                    </div>
-                  </div>
-                </Link>
-
-                <Link to="/profile">
-                  <div className="profile">
-                    <div>
-                      <svg
-                        width={26}
-                        height={26}
-                        viewBox="0 0 24 24"
-                        aria-hidden="true"
-                        className="r-4qtqp9 r-yyyyoo r-dnmrzs r-bnwqim r-1plcrui r-lrvibr r-18jsvk2 r-lwhw9o r-cnnz9e"
-                      >
-                        <g>
-                          <path d="M5.651 19h12.698c-.337-1.8-1.023-3.21-1.945-4.19C15.318 13.65 13.838 13 12 13s-3.317.65-4.404 1.81c-.922.98-1.608 2.39-1.945 4.19zm.486-5.56C7.627 11.85 9.648 11 12 11s4.373.85 5.863 2.44c1.477 1.58 2.366 3.8 2.632 6.46l.11 1.1H3.395l.11-1.1c.266-2.66 1.155-4.88 2.632-6.46zM12 4c-1.105 0-2 .9-2 2s.895 2 2 2 2-.9 2-2-.895-2-2-2zM8 6c0-2.21 1.791-4 4-4s4 1.79 4 4-1.791 4-4 4-4-1.79-4-4z"></path>
-                        </g>
-                      </svg>
-
-                      <span>Profile</span>
-                    </div>
-                  </div>
-                </Link>
-
-                <PostModal
-                  refreshPosts={() => handleShowPostsHomePage()}
-                  setLoadingTrue={() => setLoadingTrue()}
-                  setLoadingFalse={() => setLoadingFalse()}
-                  parentCallBack={handleCallback}
-                ></PostModal>
-              </div>
-
-              <LogoutModal></LogoutModal>
-            </nav>
-          </Col>
-
+          <LeftSideNavBar
+            refreshPosts={() => handleShowPostsHomePage()}
+            setLoadingTrue={() => setLoadingTrue()}
+            setLoadingFalse={() => setLoadingFalse()}
+            parentCallBack={handleCallback}
+          />
+          {/* xs={12} // 0px - 576px aralığı
+            sm={12} // 576px - 768px aralığı
+            md={11} // 768px - 992px aralığı
+            lg={5} // 992px - 1400px aralığı
+            xxl={5} // 1400px ve sonrası aralığı */}
           <Col
             xs={12} // 0px - 576px aralığı
             sm={12} // 576px - 768px aralığı
             md={11} // 768px - 992px aralığı
-            lg={6} // 1200px - 1400px aralığı
-            xxl={6} // 1400px ve sonrası aralığı
+            lg={
+              windowWidth <= 1201 && windowWidth >= 992
+                ? 7
+                : windowWidth > 1201
+                ? 5
+                : ""
+            } // 992px - 1400px aralığı
+            xxl={5} // 1400px ve sonrası aralığı
             className={`main-column `}
             style={{
               border: "1px solid rgba(0, 0, 0, 0.1)",
               borderTop: "none",
               borderBottom: "none",
               padding: "0px",
+              position: "relative",
             }}
           >
             <div
@@ -800,7 +730,11 @@ function MainPage() {
                     width={40}
                     height={40}
                     alt=""
-                    style={{ position: "relative", bottom: "30px" }}
+                    style={{
+                      position: "relative",
+                      bottom: "30px",
+                      borderRadius: "50%",
+                    }}
                   />
                 ) : (
                   <div>
@@ -811,7 +745,11 @@ function MainPage() {
                       fill="rgb(83, 100, 113)"
                       className="bi bi-person-circle"
                       viewBox="0 0 16 16"
-                      style={{ position: "relative", bottom: "30px" }}
+                      style={{
+                        position: "relative",
+                        bottom: "30px",
+                        borderRadius: "50%",
+                      }}
                       onClick={() =>
                         document.getElementById("formuploadModal").click()
                       }
@@ -1039,7 +977,14 @@ function MainPage() {
                         <>
                           <div key={post._id}>
                             {post.deactivatedOwner ? null : (
-                              <div className="each-post" key={post._id}>
+                              <div
+                                onClick={() => {
+                                  console.log("Post box parent class =>", post);
+                                  setclickedPostBox(post);
+                                }}
+                                className="each-post"
+                                key={post._id}
+                              >
                                 <div
                                   style={{
                                     textDecoration: "none",
@@ -1710,6 +1655,14 @@ function MainPage() {
                                   {/* new version favorite repost comment finish to check */}
                                 </div>
                                 <div
+                                  onClick={() => {
+                                    console.log(
+                                      "Post box child class =>",
+                                      post
+                                    );
+                                    setclickedPostBox(post);
+                                  }}
+                                  className="border-extra"
                                   style={{
                                     borderBottom: "1px solid rgba(0,0,0,0.1)",
                                   }}
@@ -1797,7 +1750,14 @@ function MainPage() {
                         .slice(0, visibleFollowingTweets)
                         .map((post, index) => (
                           <>
-                            <div className="each-post" key={post._id}>
+                            <div
+                              onClick={() => {
+                                console.log("Post box parent class =>", post);
+                                setclickedPostBox(post);
+                              }}
+                              className="each-post"
+                              key={post._id}
+                            >
                               <div
                                 style={{
                                   textDecoration: "none",
@@ -2450,6 +2410,11 @@ function MainPage() {
                                 {/* new version favorite repost comment finish to check */}
                               </div>
                               <div
+                                onClick={() => {
+                                  console.log("Post box child class =>", post);
+                                  setclickedPostBox(post);
+                                }}
+                                className="border-extra"
                                 style={{
                                   borderBottom: "1px solid rgba(0,0,0,0.1)",
                                 }}
@@ -2530,24 +2495,8 @@ function MainPage() {
             </div>
           </Col>
           {/* finish to check  main column */}
-
           {/* 3.column burası olucak */}
-          <Col
-            className="side-bar-column d-none d-lg-block d-xxl-block"
-            xs={12} // 0px - 576px aralığı
-            sm={12} // 576px - 768px aralığı
-            md={6} // 768px - 992px aralığı
-            lg={3} // 1200px - 1400px aralığı
-            xxl={3} // 1400px ve sonrası aralığı
-            style={
-              {
-                // height: "100%",
-                // backgroundColor: "indianred",
-              }
-            }
-          >
-            Side bar column
-          </Col>
+          <RightSideColumn />
         </Row>
       </Container>
     </>
